// template/react-starter-kit/src/.generated/api/model/holidayItemDto.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Nest Starter Kit
 * NestJS + MikroORM Starter Kit API
 * OpenAPI spec version: 1.0.0
 */
import type { HolidayType } from './holidayType';

export interface HolidayItemDto {
  /** 공휴일 날짜 (YYYY-MM-DD) */
  date: string;
  /** 공휴일/휴무 명칭 */
  name: string;
  /** 공휴일 구분 (STATUTORY: 법정공휴일, CUSTOM: 특별지정휴일) */
  type: HolidayType;
}